% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoadditive.R
\docType{package}
\name{geoaddSAE2-package}
\alias{geoaddSAE2}
\alias{geoaddSAE2-package}
\title{geoaddSAE2: Geoadditive Small Area Estimation for Area Level}
\description{
Fits area-level Geoadditive Small Area Estimation (SAE) models: a extension of the Fay-Herriot model that represents nonlinear covariate effects with and spatial variation, estimated by REML within a linear mixed model framework (via \pkg{mgcv}), with Mean Squared Error obtained by parametric bootstrap. The package also lets users compare the geoadditive model against the classical Fay-Herriot and Spatial Fay-Herriot models, both of which are provided by, and fitted with, the existing \pkg{sae} package.
}
\details{
Main entry point: \code{\link{geosae}}.
}
\author{
\strong{Maintainer}: Amalyah Rizky Khairunnisa Sutarto \email{222212489@stis.ac.id}

Authors:
\itemize{
  \item Amalyah Rizky Khairunnisa Sutarto \email{222212489@stis.ac.id}
  \item Novi Hidayat Pusponegoro \email{novie@stis.ac.id} [thesis advisor]
}

}
\keyword{internal}
