% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoadditive.R
\name{print.geosae}
\alias{print.geosae}
\title{Print method for geosae object}
\usage{
\method{print}{geosae}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{geosae}}

\item{digits}{Number of decimal digits to print. Default 4.}

\item{...}{Additional arguments}
}
\description{
Print method for geosae object
}
