% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoadditive.R
\name{summary.geosae}
\alias{summary.geosae}
\title{Summary method for geosae object}
\usage{
\method{summary}{geosae}(object, ...)
}
\arguments{
\item{object}{An object of class \code{geosae}}

\item{...}{Additional arguments}
}
\description{
Summary method for geosae object
}
